Camera can be moved and placed anywhere. Composition object created that allows the creation of more complex objects using triangles.

+++ src/Material.kt
@@ -1,3 +1,3 @@
-data class Material (val albedo: Point, var lambertian: Float, val refl: Float) {
-    constructor() : this(Point(1f,1f,1f), 1f, 0f)
+data class Material (val albedo: Point, var lambertian: Double, val refl: Double) {
+    constructor() : this(Point(1.0,1.0,1.0), 1.0, 0.0)
 }
